dashboard/pyproject.toml
[tool.poetry]
name = "dashboard"
version = "0.1.0"
description = ""
authors = ["Timon Schmelzer <timon.schmelzer@grandcentrix.net>"]

[tool.poetry.dependencies]
python = "^3.9"
streamlit = "1.0.0"
pandas = "^1.3.2"
ipython = "^7.27.0"
plotly = "^5.3.1"
watchdog = "^2.1.5"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
